% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidro_metadata.R
\docType{data}
\name{hidro_metadata}
\alias{hidro_metadata}
\title{Metadata of hydrological stations on Hidroweb database}
\format{A data frame with 46036 rows and 13 variables:
\describe{
\item{Codigo}{station code ID}
\item{Longitude}{degrees}
\item{Latitude}{degrees}
\item{Altitude}{masl}
\item{AreaDrenagem}{km^-2^}
\item{Nome}{station name}
...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
hidro_metadata
}
\description{
A dataset containing coordinates and additional attributes of 46036
 hydrological stations
}
\keyword{datasets}
