% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidroweb_meta.R
\name{hidroweb_meta}
\alias{hidroweb_meta}
\title{Get metadata of hydrological stations from Hidroweb database}
\usage{
hidroweb_meta()
}
\value{
data frame with station metadata
}
\description{
Get metadata of hydrological stations from Hidroweb database
}
