% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidrowebdown-package.R
\docType{package}
\name{hidrowebdown-package}
\alias{hidrowebdown}
\alias{hidrowebdown-package}
\title{hidrowebdown: Download hydrological data from Hidroweb-ANA.}
\description{
Downloading data from the major database of hydrological data in Brazil ([Hidroweb](http: //hidroweb.ana.gov.br)-National Water Agency ([ANA](www.ana.gov.br)) from R.
}
\author{
\strong{Maintainer}: Jonatan Tatsch \email{jdtatsch@gmail.com}

}
\keyword{internal}
