% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidrowebdown-package.R
\docType{package}
\name{hidrowebdown-package}
\alias{hidrowebdown}
\alias{hidrowebdown-package}
\title{hidrowebdown: Download hydrological data from Hidroweb-ANA}
\description{
Downloading data from the major database of
hydrological data in Brazil from R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lhmet/hidrowebdown}
}

}
\author{
\strong{Maintainer}: Jonatan Tatsch \email{jdtatsch@gmail.com}

}
\keyword{internal}
